Move bulk import to lib
