+++ backend/app/lib/bulk_import/bulk_importer.rb
@@ -8,7 +8,7 @@ require_relative "digital_object_handler"
 require_relative "lang_handler"
 require_relative "notes_handler"
 require_relative "subject_handler"
-require_relative "../../../lib/uri_resolver"
+require_relative "../../lib/uri_resolver"
 require "nokogiri"
 require "rubyXL"
 require "asutils"
