+++ backend/app/lib/bulk_import/container_instance_handler.rb
