+++ backend/app/lib/bulk_import/cv_list.rb
@@ -1,4 +1,4 @@
-require_relative "../../../model/enumeration"
+require_relative "../../model/enumeration"
 require_relative "crud_helpers"
 require_relative "bulk_import_mixins"
